+++ tests/Integration/AbstractEndToEndTest.php
@@ -1,0 +1,108 @@
+<?php
+
+declare(strict_types=1);
+
+namespace webignition\BasilCliCompiler\Tests\Integration;
+
+use PHPUnit\Framework\TestCase;
+use Symfony\Component\Yaml\Yaml;
+use webignition\BaseBasilTestCase\AbstractBaseTest;
+use webignition\BasilCliCompiler\Tests\DataProvider\FixturePaths;
+use webignition\BasilCliCompiler\Tests\Model\CliArguments;
+use webignition\BasilCliCompiler\Tests\Model\CompilationOutput;
+use webignition\BasilCliCompiler\Tests\Model\ExpectedGeneratedTest;
+use webignition\BasilCliCompiler\Tests\Model\ExpectedGeneratedTestCollection;
+use webignition\BasilCliCompiler\Tests\Services\ClassNameReplacer;
+use webignition\BasilCompilerModels\SuiteManifest;
+
+abstract class AbstractEndToEndTest extends TestCase
+{
+    private ClassNameReplacer $classNameReplacer;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->classNameReplacer = new ClassNameReplacer();
+    }
+
+    protected function tearDown(): void
+    {
+        parent::tearDown();
+
+        $directoryIterator = new \DirectoryIterator(FixturePaths::getTarget());
+
+        foreach ($directoryIterator as $item) {
+            /** @var \DirectoryIterator $item */
+            if ('php' === $item->getExtension() && $item->isFile() && $item->isWritable()) {
+                unlink($item->getPathname());
+            }
+        }
+    }
+
+    /**
+     * @dataProvider generateDataProvider
+     */
+    public function testGenerate(
+        CliArguments $cliArguments,
+        ExpectedGeneratedTestCollection $expectedGeneratedTests
+    ): void {
+        $compilationOutput = $this->getCompilationOutput($cliArguments);
+        $this->assertSame(0, $compilationOutput->getExitCode());
+
+        $suiteManifest = SuiteManifest::fromArray((array) Yaml::parse($compilationOutput->getContent()));
+
+        $suiteManifestConfiguration = $suiteManifest->getConfiguration();
+        self::assertSame($cliArguments->getSource(), $suiteManifestConfiguration->getSource());
+        self::assertSame($cliArguments->getTarget(), $suiteManifestConfiguration->getTarget());
+        self::assertSame(AbstractBaseTest::class, $suiteManifestConfiguration->getBaseClass());
+
+        $testManifests = $suiteManifest->getTestManifests();
+        self::assertCount(count($expectedGeneratedTests), $testManifests);
+
+        $localTarget = getcwd() . FixturePaths::TARGET;
+
+        foreach ($testManifests as $index => $testManifest) {
+            $testPath = $testManifest->getTarget();
+            $localTestPath = str_replace($cliArguments->getTarget(), $localTarget, $testPath);
+            self::assertFileExists($localTestPath);
+
+            $expectedGeneratedTest = $expectedGeneratedTests[$index];
+            $generatedTestContent = (string) file_get_contents($localTestPath);
+
+            $generatedTestContent = $this->classNameReplacer->replaceNamesInContent(
+                $generatedTestContent,
+                [$expectedGeneratedTest->getReplacementClassName()]
+            );
+
+            $this->assertSame($expectedGeneratedTest->getExpectedContent(), $generatedTestContent);
+        }
+    }
+
+    /**
+     * @return array[]
+     */
+    public function generateDataProvider(): array
+    {
+        return [
+            'single test' => [
+                'cliArguments' => new CliArguments(
+                    $this->getRemoteSourcePrefix() . '/Test/example.com.verify-open-literal.yml',
+                    $this->getRemoteTarget(),
+                ),
+                'expectedGeneratedTests' => new ExpectedGeneratedTestCollection([
+                    new ExpectedGeneratedTest(
+                        'GeneratedVerifyOpenLiteralChrome',
+                        '/tests/Fixtures/php/Test/GeneratedVerifyOpenLiteralChrome.php',
+                    ),
+                ]),
+            ],
+        ];
+    }
+
+    abstract protected function getRemoteSourcePrefix(): string;
+
+    abstract protected function getRemoteTarget(): string;
+
+    abstract protected function getCompilationOutput(CliArguments $cliArguments): CompilationOutput;
+}
