+++ tests/Integration/Image/CompilerTest.php
@@ -4,91 +4,25 @@ declare(strict_types=1);
 
 namespace webignition\BasilCliCompiler\Tests\Integration\Image;
 
-use PHPUnit\Framework\TestCase;
-use Symfony\Component\Yaml\Yaml;
-use webignition\BaseBasilTestCase\AbstractBaseTest;
-use webignition\BasilCliCompiler\Tests\DataProvider\FixturePaths;
+use webignition\BasilCliCompiler\Tests\Integration\AbstractEndToEndTest;
 use webignition\BasilCliCompiler\Tests\Model\CliArguments;
 use webignition\BasilCliCompiler\Tests\Model\CompilationOutput;
-use webignition\BasilCliCompiler\Tests\Model\ExpectedGeneratedTest;
-use webignition\BasilCliCompiler\Tests\Model\ExpectedGeneratedTestCollection;
-use webignition\BasilCliCompiler\Tests\Services\ClassNameReplacer;
-use webignition\BasilCompilerModels\SuiteManifest;
 use webignition\TcpCliProxyClient\Client;
 use webignition\TcpCliProxyClient\HandlerFactory;
 
-class CompilerTest extends TestCase
+class CompilerTest extends AbstractEndToEndTest
 {
-    private ClassNameReplacer $classNameReplacer;
-
-    protected function setUp(): void
+    protected function getRemoteSourcePrefix(): string
     {
-        parent::setUp();
-
-        $this->classNameReplacer = new ClassNameReplacer();
+        return '/app/source';
     }
 
-    /**
-     * @dataProvider generateDataProvider
-     */
-    public function testGenerate(
-        CliArguments $cliArguments,
-        ExpectedGeneratedTestCollection $expectedGeneratedTests
-    ): void {
-        $compilationOutput = $this->getCompilationOutput($cliArguments);
-        $this->assertSame(0, $compilationOutput->getExitCode());
-
-        $suiteManifest = SuiteManifest::fromArray((array) Yaml::parse($compilationOutput->getContent()));
-
-        $suiteManifestConfiguration = $suiteManifest->getConfiguration();
-        self::assertSame($cliArguments->getSource(), $suiteManifestConfiguration->getSource());
-        self::assertSame($cliArguments->getTarget(), $suiteManifestConfiguration->getTarget());
-        self::assertSame(AbstractBaseTest::class, $suiteManifestConfiguration->getBaseClass());
-
-        $testManifests = $suiteManifest->getTestManifests();
-        self::assertCount(count($expectedGeneratedTests), $testManifests);
-
-        $localTarget = getcwd() . FixturePaths::TARGET;
-
-        foreach ($testManifests as $index => $testManifest) {
-            $testPath = $testManifest->getTarget();
-            $localTestPath = str_replace($cliArguments->getTarget(), $localTarget, $testPath);
-            self::assertFileExists($localTestPath);
-
-            $expectedGeneratedTest = $expectedGeneratedTests[$index];
-            $generatedTestContent = (string) file_get_contents($localTestPath);
-
-            $generatedTestContent = $this->classNameReplacer->replaceNamesInContent(
-                $generatedTestContent,
-                [$expectedGeneratedTest->getReplacementClassName()]
-            );
-
-            $this->assertSame($expectedGeneratedTest->getExpectedContent(), $generatedTestContent);
-        }
-    }
-
-    /**
-     * @return array[]
-     */
-    public function generateDataProvider(): array
+    protected function getRemoteTarget(): string
     {
-        return [
-            'single test' => [
-                'cliArguments' => new CliArguments(
-                    '/app/source/Test/example.com.verify-open-literal.yml',
-                    '/app/tests'
-                ),
-                'expectedGeneratedTests' => new ExpectedGeneratedTestCollection([
-                    new ExpectedGeneratedTest(
-                        'GeneratedVerifyOpenLiteralChrome',
-                        '/tests/Fixtures/php/Test/GeneratedVerifyOpenLiteralChrome.php',
-                    ),
-                ]),
-            ],
-        ];
+        return '/app/tests';
     }
 
-    private function getCompilationOutput(CliArguments $cliArguments): CompilationOutput
+    protected function getCompilationOutput(CliArguments $cliArguments): CompilationOutput
     {
         $output = '';
         $exitCode = 0;
